feat: Weather App UI and Data Management Updates
## UserDelegate

Added new properties and methods to enhance weather display functionality:

- weatherHighlights: [Int] - Store highlighted weather conditions
- displayPastHours: bool - Toggle historical weather data visibility
- multiColorSymbols: bool - Enable multi-colored weather symbols

Added methods for:

- Temperature management (updateTempMeasurement, toggleTempMeasurement)
- Display preferences (toggleDisplayPastHours, toggleMultiColorSymbols)
- Weather highlights (decodeHighlights, saveHighlights)
- Location management (deleteLocation)

# UserDelegateAttribute

Added new properties to display custom information to the user

- highlights: [Int] - Store highlighted weather conditions
- displaysPastHours: Bool - Toggle historical weather data visibility
- multiColorSymbols: Bool - Enable multi-colored weather symbols

# Day

Updated property types

- uv changed from an Int to a Double

# Hour

Updated property types

- uv changed from an Int to a Double

# HomeTabView

Added HomeTabView: A generic SwiftUI view that manages tab navigation with loading states, error handling, and dynamic content rendering based on ViewState

# HomeView

Reconfigured the layout of the view to utilize the HomeViewTab

### **SavedLocations**

Added view to display saved locations by the user

### **SettingsView**

Added tempMeasurmentToggle, tempMeasurement, displayPastHours as properties of settings view and views to update their values

### **WeatherHighlightSettings**

Added view to allow the user to select their top three weather highlights to view on their homepage

+++ Weather Provider/Model/Weather Network/Network Results/Day.swift
@@ -14,7 +14,7 @@ struct Day: Codable {
     var totalsnowCM, avgvisKM, avgvisMiles, avghumidity: Double
     var dailyWillItRain, dailyChanceOfRain, dailyWillItSnow, dailyChanceOfSnow: Int
     var condition: Condition
-    var uv: Int
+    var uv: Double
     
     enum CodingKeys: String, CodingKey {
         case maxtempC = "maxtemp_c"
